move expression to separate module
