+++ app/domain/expressions/tokenizer.py
@@ -2,7 +2,7 @@ import enum
 import re
 from dataclasses import dataclass
 
-# Use https://regex101.com/ to explain this regex
+# Use https://regex101.com/ to explain these regex
 function_call_pattern = r'^([a-z0-9-]+):((?:[a-zA-Z0-9_+.]+)|(?:"[^"]+"))'
 operator_pattern = r"^([a-z-]+)\s+"
